# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for The Soraa Cafe website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuCategory {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  itemCount: number;
}

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  categoryId: number;
  categoryName: string;
  isVeg: boolean;
  isPopular: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface GalleryPhoto {
  id: number;
  imageUrl: string;
  title: string;
  /** @nullable */
  description?: string | null;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  createdAt: string;
}

export interface OpeningHour {
  day: string;
  hours: string;
}

export interface AboutInfo {
  name: string;
  tagline: string;
  description: string;
  address: string;
  phone: string;
  email: string;
  openingHours: OpeningHour[];
}

export interface ErrorResponse {
  error: string;
}

export type GetMenuItemsParams = {
/**
 * Filter by category slug
 */
category?: string;
};