# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for The Soraa Cafe website
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/items:
    get:
      operationId: getMenuItems
      tags: [menu]
      summary: Get all menu items, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category slug
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
...
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    MenuItem:
      type: object
      required: [id, name, price, categoryId, categoryName, isVeg, isPopular]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        isVeg:
          type: boolean
        isPopular:
          type: boolean
        imageUrl:
          type: ["string", "null"]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string