# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for The Soraa Cafe website
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  categoryId: number;
  categoryName: string;
  isVeg: boolean;
  isPopular: boolean;
  /** @nullable */
  imageUrl?: string | null;
}